introduce the galerkin.Trajectory class
the class takes care of all necessary computations wrt. to the
trajectory (sigma heuristic, membership matrix) but also allows for
instatiation with precomputed data.
the core functions are still accessible outside of the class, so the
latter is just a convenience wrapper.

this fixes #14

i am not quite happy yet with the seperation of attributes (e.g. sqd)
and properties (e.g. membership).