fix wrong construction of constant eigenvector in schurvects
in the case of irreversible matrices the first vector returned by scipy.schur was not necessarily constant. overwriting with 1 thus changed the subspace
this is alleviated by constructing the unit vector from all vectors.
this is now done in the inner_simplex_algorithm() (since its the only place where this plays a role)
todo: check if isa makes assumptions on orthogonality of the vectors, since this is broken now...